Split up types module of the trigger generator into into more specific parts
This also makes modifications to support further changes to the trigger
generation that move selection of IDs to update into SIR. Only PKs of
releavant tables are selected and passed to the message queue.
